TS Move common types to core package #9779
Moved Form items to core.
