+++ modules/lib/lib-content/src/main/resources/lib/xp/content.ts
@@ -22,6 +22,7 @@ import type {
     ByteSource,
     Content,
     Filter,
+    FormItem,
     Highlight,
     HighlightResult,
     PublishInfo,
@@ -97,6 +98,12 @@ export type {
     IdsFilter,
     BooleanFilter,
     Filter,
+    FormItemSet,
+    FormItemLayout,
+    FormItemInput,
+    FormItemOptionSet,
+    FormItemInlineMixin,
+    FormItem,
 } from '@enonic-types/core';
 
 type Attachments = Content['attachments'];
@@ -140,119 +147,6 @@ interface GetAttachmentsHandler {
 
     execute(): Attachments | null;
 }
-
-export type FormItemType = 'Input' | 'ItemSet' | 'Layout' | 'OptionSet';
-
-export type InputType =
-    | 'Time'
-    | 'DateTime'
-    | 'CheckBox'
-    | 'ComboBox'
-    | 'Long'
-    | 'Double'
-    | 'RadioButton'
-    | 'TextArea'
-    | 'ContentTypeFilter'
-    | 'GeoPoint'
-    | 'TextLine'
-    | 'Tag'
-    | 'CustomSelector'
-    | 'AttachmentUploader'
-    | 'ContentSelector'
-    | 'MediaSelector'
-    | 'ImageSelector'
-    | 'Date'
-    | 'MediaUploader'
-    | 'SiteConfigurator'
-    | 'HtmlArea';
-
-export interface FormItemSet {
-    formItemType: string | FormItemType;
-    name: string;
-    label: string;
-    customText: string;
-    helpText: string;
-    maximize: boolean;
-    inputType: InputType;
-    occurrences: {
-        maximum: number;
-        minimum: number;
-    };
-    items: FormItem[];
-}
-
-export interface FormItemLayout {
-    formItemType: string | FormItemType;
-    name: string;
-    label: string;
-    items: FormItem[];
-}
-
-export type ValueType =
-    | 'BinaryReference'
-    | 'Boolean'
-    | 'DateTime'
-    | 'Double'
-    | 'GeoPoint'
-    | 'Link'
-    | 'LocalDateTime'
-    | 'LocalDate'
-    | 'LocalTime'
-    | 'Long'
-    | 'PropertySet'
-    | 'Reference'
-    | 'String'
-    | 'Xml';
-
-export interface FormItemInput {
-    formItemType: string | FormItemType;
-    name: string;
-    label: string;
-    customText: string;
-    helpText: string;
-    validationRegexp: string;
-    maximize: boolean;
-    inputType: InputType;
-    occurrences: {
-        maximum: number;
-        minimum: number;
-    };
-    default: {
-        value: string;
-        type: ValueType;
-    }
-    config: {
-        [configName: string]: {
-            [attributeKey: string]: string;
-            value: string;
-        }[]
-    }
-}
-
-export interface FormItemOptionSet {
-    formItemType: string | FormItemType;
-    name: string;
-    label: string;
-    expanded: boolean;
-    helpText: string;
-    occurrences: {
-        maximum: number;
-        minimum: number;
-    };
-    selection: {
-        maximum: number;
-        minimum: number;
-    };
-    options: {
-        name: string;
-        label: string;
-        helpText: string;
-        default: boolean;
-        items: FormItem[]
-    }[];
-}
-
-export type FormItem = FormItemSet | FormItemLayout | FormItemOptionSet | FormItemInput;
 
 /**
  * This function fetches a content.
